Remove gk_geometry_fromfile.c and move gk_geometry_new into gk_geometry.c . Also fix bcast_rank in apps/gyrokinetic based on a small change to logic in 663abf2
