--- zero/gk_geometry_fromfile.c
@@ -1,62 +1,0 @@
-#include <gkyl_alloc.h>
-#include <gkyl_alloc_flags_priv.h>
-#include <gkyl_array.h>
-#include <gkyl_array_ops.h>
-#include <gkyl_array_rio.h>
-#include <gkyl_basis.h>
-#include <gkyl_calc_derived_geo.h>
-#include <gkyl_calc_metric.h>
-#include <gkyl_eval_on_nodes.h>
-#include <gkyl_gk_geometry.h>
-#include <gkyl_math.h>
-#include <gkyl_nodal_ops.h>
-
-
-struct gk_geometry*
-gkyl_gk_geometry_new(struct gk_geometry* geo_host, struct gkyl_gk_geometry_inp *geometry_inp, bool use_gpu)
-{
-
-#ifdef GKYL_HAVE_CUDA
-  if(use_gpu) {
-    return gkyl_gk_geometry_cu_dev_new(geo_host, geometry_inp);
-  } 
-#endif 
-
-  struct gk_geometry *up = gkyl_malloc(sizeof(struct gk_geometry));
-  up->basis = geometry_inp->basis;
-  up->local = geometry_inp->local;
-  up->local_ext = geometry_inp->local_ext;
-  up->global = geometry_inp->global;
-  up->global_ext = geometry_inp->global_ext;
-  up->grid = geometry_inp->grid;
-
-  // bmag, metrics and derived geo quantities
-  up->mc2p = gkyl_array_new(GKYL_DOUBLE, 3*up->basis.num_basis, up->local_ext.volume);
-  up->bmag = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->g_ij = gkyl_array_new(GKYL_DOUBLE, 6*up->basis.num_basis, up->local_ext.volume);
-  up->dxdz = gkyl_array_new(GKYL_DOUBLE, 9*up->basis.num_basis, up->local_ext.volume);
-  up->dzdx = gkyl_array_new(GKYL_DOUBLE, 9*up->basis.num_basis, up->local_ext.volume);
-  up->jacobgeo = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->jacobgeo_inv = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->gij = gkyl_array_new(GKYL_DOUBLE, 6*up->basis.num_basis, up->local_ext.volume);
-  up->b_i = gkyl_array_new(GKYL_DOUBLE, 3*up->basis.num_basis, up->local_ext.volume);
-  up->cmag = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->jacobtot = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->jacobtot_inv = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->bmag_inv = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->bmag_inv_sq = gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->gxxj= gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->gxyj= gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->gyyj= gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->gxzj= gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->eps2= gkyl_array_new(GKYL_DOUBLE, up->basis.num_basis, up->local_ext.volume);
-  up->bmag_mid = gkyl_array_new(GKYL_DOUBLE, 1, 1);
-
-  up->flags = 0;
-  GKYL_CLEAR_CU_ALLOC(up->flags);
-  up->ref_count = gkyl_ref_count_init(gkyl_gk_geometry_free);
-  up->on_dev = up; // CPU eqn obj points to itself
-                   
-  return up;
-}
-
